backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cohere==5.13.0
cryptography==46.0.4
deprecation==2.1.0
dnspython==2.8.0
elevenlabs==2.0.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastavro==1.12.1
filelock==3.20.3
fsspec==2026.1.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.3.5
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
load-dotenv==0.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==26.0
parameterized==0.9.0
postgrest==2.27.2
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
realtime==2.27.2
requests==2.32.5
rich==14.3.1
rich-toolkit==0.17.2
rignore==0.7.6
sentry-sdk==2.51.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
sse-starlette==2.2.1
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
tokenizers==0.22.2
tqdm==4.67.2
typer==0.21.1
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
elevenlabs>=2.0.0
